refactor: public 파일/디렉토리 명 대소문자 업데이트 - logIn, logOut -> login, logout으로 변경
