+++ apps/client/public/images/logout/index.ts
